+++ demo/lib/examples/invoice.dart
@@ -451,7 +451,7 @@ class Invoice {
       'Total'
     ];
 
-    return pw.Table.fromTextArray(
+    return pw.TableHelper.fromTextArray(
       border: null,
       cellAlignment: pw.Alignment.centerLeft,
       headerDecoration: pw.BoxDecoration(