+++ printing/ios/Classes/PrintJob.swift
@@ -32,7 +32,7 @@ public class PrintJob: UIPrintPageRenderer, UIPrintInteractionControllerDelegate
     private let semaphore = DispatchSemaphore(value: 0)
     private var dynamic = false
     private var currentSize: CGSize?
-    
+
     public init(printing: PrintingPlugin, index: Int) {
         self.printing = printing
         self.index = index
@@ -134,13 +134,15 @@ public class PrintJob: UIPrintPageRenderer, UIPrintInteractionControllerDelegate
 
         printing.onCompleted(printJob: self, completed: completed, error: error?.localizedDescription as NSString?)
     }
-
-    public func printInteractionController(_ printInteractionController: UIPrintInteractionController, cutLengthFor paper: UIPrintPaper) -> CGFloat {
-        if currentSize == nil{
-            return  paper.paperSize.height
+    
+    public func printInteractionController(_ printController: UIPrintInteractionController, choosePaper paperList: [UIPrintPaper]) -> UIPrintPaper {
+        if currentSize == nil {
+            return paperList[0]
         }
-
-        return currentSize!.height
+        
+        let bestPaper = UIPrintPaper.bestPaper(forPageSize: currentSize!, withPapersFrom: paperList)
+        
+        return bestPaper
     }
 
     func printPdf(name: String, withPageSize size: CGSize, andMargin margin: CGRect, withPrinter printerID: String?, dynamically dyn: Bool) {
@@ -312,18 +314,19 @@ public class PrintJob: UIPrintPageRenderer, UIPrintInteractionControllerDelegate
     }
 
     public func rasterPdf(data: Data, pages: [Int]?, scale: CGFloat) {
-        let provider = CGDataProvider(data: data as CFData)!
-        let document = CGPDFDocument(provider)
-        if document == nil {
+        guard
+            let provider = CGDataProvider(data: data as CFData),
+            let document = CGPDFDocument(provider)
+        else {
             printing.onPageRasterEnd(printJob: self, error: "Cannot raster a malformed PDF file")
             return
         }
 
         DispatchQueue.global().async {
-            let pageCount = document!.numberOfPages
+            let pageCount = document.numberOfPages
 
             for pageNum in pages ?? Array(0 ... pageCount - 1) {
-                guard let page = document!.page(at: pageNum + 1) else { continue }
+                guard let page = document.page(at: pageNum + 1) else { continue }
                 let angle = CGFloat(page.rotationAngle) * CGFloat.pi / -180
                 let rect = page.getBoxRect(.mediaBox)
                 let width = Int(abs((cos(angle) * rect.width + sin(angle) * rect.height) * scale))