+++ printing/macos/Classes/PrintJob.swift
@@ -252,18 +252,19 @@ public class PrintJob: NSView, NSSharingServicePickerDelegate {
     }
 
     public func rasterPdf(data: Data, pages: [Int]?, scale: CGFloat) {
-        let provider = CGDataProvider(data: data as CFData)!
-        let document = CGPDFDocument(provider)
-        if document == nil {
+        guard
+            let provider = CGDataProvider(data: data as CFData),
+            let document = CGPDFDocument(provider)
+        else {
             printing.onPageRasterEnd(printJob: self, error: "Cannot raster a malformed PDF file")
             return
         }
 
         DispatchQueue.global().async {
-            let pageCount = document!.numberOfPages
+            let pageCount = document.numberOfPages
 
             for pageNum in pages ?? Array(0 ... pageCount - 1) {
-                guard let page = document!.page(at: pageNum + 1) else { continue }
+                guard let page = document.page(at: pageNum + 1) else { continue }
                 let angle = CGFloat(page.rotationAngle) * CGFloat.pi / -180
                 let rect = page.getBoxRect(.mediaBox)
                 let width = Int(abs((cos(angle) * rect.width + sin(angle) * rect.height) * scale))